requirements.txt
faiss-cpu>=1.9.0
sentence-transformers>=3.0.0
scikit-learn>=1.5.0
streamlit>=1.35.0
PyMuPDF>=1.24.3
requests>=2.31.0
numpy>=2.0.0
python-dotenv>=1.0.1